make data module private
