+++ mesmer/volc.py
@@ -2,10 +2,31 @@ import warnings
 
 import xarray as xr
 
-from mesmer._core._data import load_stratospheric_aerosol_optical_depth_obs
+from mesmer._core._data import _load_aod_obs
 from mesmer._core.utils import _assert_annual_data, _check_dataarray_form
 from mesmer.datatree import _datatree_wrapper
 from mesmer.stats import LinearRegression
+
+
+def load_stratospheric_aerosol_optical_depth_obs(*, version="2022", resample=True):
+    """load stratospheric aerosol optical depth data - a proxy for volcanic activity
+
+    Parameters
+    ----------
+    version : str, default: "2022"
+        Which version of the dataset to load. Currently only "2022" is available.
+    resample : bool, default: True
+        Whether to resample the data to annual resolution.
+
+    Returns
+    -------
+    stratospheric_aerosol_optical_depth_obs : xr.DataArray
+        DataArray of stratospheric aerosol optical depth observations.
+    """
+
+    aod = _load_aod_obs(version=version, resample=resample)
+
+    return aod.copy()
 
 
 def _load_and_align_strat_aod_obs(
